#3054 - Email notification for blocked disbursement to ministry only once (#4245)
- Created a base class for e-Cert notifications (`ECertNotification`)
and segregated the logic to generate blocked notifications into
`MinistryBlockedDisbursementNotification` and
`StudentBlockedDisbursementNotification`.
- Existing logic was moved into the new files.
- Code for the method `shouldCreateNotification` for the Students was
moved and no changes were done for this PR, unless critical no comments
will be addressed.
- `ECertNotification` can be used for any other e-Cert notification as
required, not only blocked disbursement.

_Please note that the student data required to generate the notification
is now retrieved twice in the event that the disbursement is blocked and
the Ministry should be notified for the first time. The benefit of
having the separation between the logic for the different notifications
seems greater than an occasional extra query to the student's table.
Please note that the query was simplified from what it used to be._

_Please also note that the below technical AC was not executed. The
queries have different approaches, sometimes checking the existence,
sometimes requiring a count, and sometimes requiring a count and a date
last sent. I did not see a benefit for this PR. Please share an
opinion/suggestion if needed._
```
Create a generic method to check, if the email notifications are already sent to a particular template and metadata( disbursementId).
```