#4253 - Upgrade Redis Version and increase Memory - Part 1 (#4292)
Team, the below files are copied from bitnami
(https://github.com/bitnami/charts/tree/main/bitnami/redis-cluster) for
the charts, so please ignore the review on these files below

![image](https://github.com/user-attachments/assets/20eba934-660b-45ab-bcd4-feb629cb72cf)
or in folders 
> base files in devops/helm/redis-cluster
> devops/helm/redis-cluster/templates

Make file is created for easy installtion of helm charts.

The values.yaml overridden with values from the values-{NAMESPACE}.yaml.
The recommendation values are not given in the
https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/
link given, but as per the examples, redis cluster starts at a minimum
configuration of Nano where the values are given in this
https://github.com/bitnami/charts/blob/main/bitnami/common/templates/_resources.tpl#L15.

I will create a the next PR with the actual values for namespaces
values-{NAMESPACE}.yaml.

To enable the github action initial commit of the file is created.

NOTE: Additional file changes
sources/camunda-docker-compose-core.yml - missed network type
sources/packages/forms/Dockerfile.dev is changed to enable development
environment login