#3254 - CAS AP Invoice - Invoices and Batches - Scheduler (#4297)
- Created the new scheduler default setup including DB migrations.
- Added information in the wiki with the agreed time with the business.
- Created only the first E2E to allow some testing. Adding further E2E
is planned but it will also demand adding more factories hence this was
a point to stop for now.
- Created DB seeding for the new CAS distribution accounts to allow easy
execution assertions on expected distribution accounts. It may be shared
with API when API tests are created.
- _Note:_ DB seeding was never enabled for queue-consumers and few files
were modified to allow it following the same already in place for the
API.
- Minor refactors were executed in the DB seeding to allow the system
user to be retrieved and the default logger to be used instead of
`console`.
- The `sequence-control` service was changed to allow outside control of
the sequence increment.

## Pending Receipts Query

- Query to retrieve the pending receipts was kept as "query builder" to
enforce the inner joins (even if some of them would work as left joins
enforced by the where condition).
- Query to retrieve the pending receipts currently does not have a stop
date to look for pending disbursements and is just retrieving anything
that is pending.

## Saving Operations

The method to save the invoices is using a single transaction and chunk
inserts. The 150-chunk size was determined by local tests using 50 to
500-chunk sizes. The tests were realized with 1000 invoice records and
each one had 4 invoice details where it took around 1 second to complete
the DB persistence operation. In terms of SQL parameters, Postgres can
accept a maximum of [65K
parameters](https://www.postgresql.org/docs/current/limits.html) in a
single command.

The save operation in chunks was not logging any additional select and
was grouping the insert operations as below for invoices and similar for
invoice details.

```sql
INSERT INTO
    "sims"."cas_invoices"("created_at", "updated_at", "invoice_number", "invoice_status", "invoice_status_updated_on", "errors", "creator", "modifier", "cas_invoice_batch_id", "disbursement_receipt_id", "cas_supplier_id")
VALUES 
    (DEFAULT, DEFAULT, $1, $2, $3, DEFAULT, $4, DEFAULT, $5, $6, $7),
    (DEFAULT, DEFAULT, $8, $9, $10, DEFAULT, $11, DEFAULT, $12, $13, $14),
    (DEFAULT, DEFAULT, $15, $16, $17, DEFAULT, $18, DEFAULT, $19, $20, $21),
    (DEFAULT, DEFAULT, $22, $23, $24, DEFAULT, $25, DEFAULT, $26, $27, $28),
    (DEFAULT, DEFAULT, $29, $30, $31, DEFAULT, $32, DEFAULT, $33, $34, $35),
    (DEFAULT, DEFAULT, $36, $37, $38, DEFAULT, $39, DEFAULT, $40, $41, $42)
    (...)
```
## Migration revert


![image](https://github.com/user-attachments/assets/fd09ab19-eb45-43d3-91a3-19e57b661120)