#3254 - CAS AP Invoice - Invoices and Batches - UI (#4305)
New proposed UI to allow invoice batches download. The content is
completely subject to be changed and the goal is to have a starting
point to make some progress towards the final solution.

- Used a data table with server-side pagination.
- Enabled sort operation only for the batch date for now.
- Enable some filters on statuses.
- Added temporary download button (functionality to be added during this
sprint in an upcoming PR).
- Added temporary modals to the UI to later add the functionality for
batch approval/rejection (it may or may not be an effort in the current
sprint).
- Final Ministry roles to be created in an upcoming PR.
- E2E tests to be added in an upcoming PR.


![image](https://github.com/user-attachments/assets/68b3cadd-3796-4752-a9cd-044e810fdcd3)
_Note: the above UI was changed during PR to remove the "Only" prefix
that can be noticed in the below images._


![image](https://github.com/user-attachments/assets/831aa603-f8b6-49f1-9a21-241231ea2cb0)


![image](https://github.com/user-attachments/assets/e0fcf603-309a-49fd-a347-dd6583dd24cf)