#3594 - Email to Student - 2nd Disbursement has passed, still pending (Part 3) (#4310)
As the final part of the story, this PR has the following changes:
- Added E2E tests for the second disbursement reminder notification.
- Should generate a second disbursement reminder notification for a
student when the application is Completed and the second disbursement
with Required COE status is still pending when its disbursement date has
passed.
- Should not generate a second disbursement reminder notification for a
student when a notification is already sent for the second disbursement
still pending.
- Should not generate a second disbursement reminder notification for a
student when an application has two disbursements with both COEs
completed but the second disbursement is still pending.


![image](https://github.com/user-attachments/assets/19f8e5e6-6655-4957-8415-6539024ccee9)