#4247 - Create Application Versioning - Part 3 (#4340)
### As a part of this PR, the following were completed:

- Adjusted the necessary controllers / methods in the API to allow the
student application view on the Ministry and the Institution to
optionally retrieve the latest application details based on the provided
application id (considered as parent application) or to retrieve the
application details of the passed application id based on the value of
the query parameter `isParentApplication` set to true or false
respectively.
- Modified the existing method to `check if a previous application
exists` and added a method to `retrieve the current application from the
parent application`.
- Adjusted the existing API `e2e tests`.

---------

Co-authored-by: Dheepak Ramanathan <dheepak.rr@aot-technologies.com>