#3739 - Release DB Migrations Rollback - Allow migration to fail on CI/CD (#4457)
When CI/CD is executing the migrations, the job was considered
successful even if a migration failed to be executed.
Changed the exceptions handling behavior when executed in the CI/CD
pipeline or using the REPL mode.

The initial raised PR approach was changed based on the below finding.
> Following further testing, it was observed that the RPEL is resilient
to unhandled exceptions. Therefore, the code was simplified by removing
the catch block, allowing exceptions to be thrown in all cases. This
resolves the CI/CD migration issue and does not affect REPL mode, while
also simplifying the codebase.