#4354 - Change Request Revamp - 1st effort - Workflow Changes (#4464)
## Workers changes

- Created a new worker `application-change-request-approval` to process
the Ministry review of an application change request.
- The worker will be executed only for applications with the edit status
set as `Change in progress`.
- It is the worker's responsibility to set the application's edit status
to `Change pending approval` to keep the workflow waiting for some new
action from the Ministry, to approve or reject it, or from the student,
to cancel it.

### E2E Tests

- ApplicationController(e2e)-applicationChangeRequestApproval
- Should update application edit status to Change pending approval when
the Ministry approval is required.
- Should not update the application edit status when it is already set
as Change pending approval.
- Should not update the application edit status when it is set as Change
cancelled.

## Workflow changes

### New alternative flow to skip application status change and
exceptions verifications

- The application status will remain as it is (`Edited`).
- No exceptions verifications will be executed.


![image](https://github.com/user-attachments/assets/ddd6b8c1-1580-4de1-b516-3f046d4bc44d)

### New `change request` approval flow

- When the application edit status is set to `Change in progress`, the
new flow will be executed and wait for some action from the Ministry or
the Student, as explained before.
- The edit status will be set as `Change pending approval` and the
workflow will be waiting for timer to check the most recent DB status or
for a message to be received (following the same approach from PIR and
exceptions).
- Also following the same approach from exceptions, if the workflow will
receive a "cancellation status" (`Change cancelled` or `Change
declined`) and it will proceed to the `end`.


![image](https://github.com/user-attachments/assets/cdf7b59a-9f00-4bad-90d4-b591f619edb5)

### PIR ignored


![image](https://github.com/user-attachments/assets/34fe2bcf-eb86-405e-b601-740e806e7b9a)

### E2E Tests
- E2E Test Workflow assessment gateway on change requests for <program
year>
- Should complete the assessment calculations when a change request was
approved by the Ministry and a message was sent to the workflow.
- Should end the assessment workflow when a change request is declined
by the Ministry via message.
- Should end the assessment workflow when a change request is declined
by the Ministry via task result.
- Should end the assessment workflow when a change request is cancelled
by the student via message.
- Should end the assessment workflow when a change request is cancelled
by the student via task result.

## Non-related changes

- VS Code tasks to execute E2E tests were updated to allow the execution
of files with the same name individually. Before, files with the same
name were executed together, for instance, the E2E files for different
program years that share the same name.
- Create two new tasks in the workspace for easy execution of Camunda
and forms definitions.

## TODO
- Set the appropriate assessment status before the workflows execution
ends, adapting and reusing the `Workflow wrap up` task or creating
something similar, where the last seems more appropriate because we just
need to execute something similar to the method
[updateAssessmentStatusAndSaveWorkflowData](https://github.com/bcgov/SIMS/blob/ac0b8ca1951210e0691b03617cdd9e6940afaabc/sources/packages/backend/apps/workers/src/services/student-assessment/student-assessment.service.ts#L297).
_Note:_ the same could be reused for when the workflow is ended due to
the Ministry declining exceptions.