#4354 - Change Request Revamp - 1st effort - Warping up workflow during "in-progress" application (#4491)
- Created a variation for the existing worker "workflow-wrap-up" to
allow its usage before the "end workflow" for the `assessment-gateway`
"Applications In Progress" lane, as shown in items 1 to 3 in the below
image (image 4 is working as before).
- The worker is backward compatible with the existing workflow, which
means that once deployed, workflows executing the previous version where
the new header is not provided will be able to finish its process as
before. The absence of the header makes it works as before.
- _Note:_ fixed an issue when a `job.complete()` was returned from
inside a transaction block with the intention of finishing the worker
task but was actually finishing only the transaction block, leaving the
rest of the worker processing to still be executed.


![image](https://github.com/user-attachments/assets/b6497b5d-a5b8-4239-8f37-96318b316b19)

The workflow task now accepts a header to define its behavior, as shown
below.


![image](https://github.com/user-attachments/assets/d691eda2-86f1-4cf8-97ae-e3e4163fd61d)


![image](https://github.com/user-attachments/assets/03030e3b-ee6d-4915-83c4-63ddffcebd44)