#4431 - Student View change request revamp - View Application Versions (#4526)
## Student Applications Summary Changes

- New `View` button added with the same link from "Financial Aid
Application";
- New `Change request` button redirecting to the current "Request a
Change" for program years before 2025-2026. __Note:_ This logic was
implemented in the API to allow server validation in the future. It is
considered temporary till all program years support the new feature and
it was implemented with simplicity in mind._
- The current `Request a change` option inside the application details
always redirects to the current appeals feature. It will be changed in
an upcoming PR.
- New `Versions` expander to allow the visualization of the previously
submitted applications. Its content is currently loaded on demand.


![image](https://github.com/user-attachments/assets/f1ff4fb2-ca43-4e54-a741-6fca5354e2ef)

## Request a Change Page Changes

- Navigation menu removed from the top nav bar.
- Application ID is provided in the URL and the application number now
rendered as a read-only text input.
_Note:_ the destiny of this page and this feature as it is, in under
review. No further time will be invested and the goal is just to keep it
working for now.


![image](https://github.com/user-attachments/assets/709cf249-7116-41de-9c2d-99da85ff40e0)


![image](https://github.com/user-attachments/assets/3be2a18d-f4ab-4f09-9715-1a99a397c861)

## Non-related changes/suggestions

As part of this PR, it was introduced the
[skeleton-loaders](https://vuetifyjs.com/en/components/skeleton-loaders/#usage)
with the intention to start using them along the solution.


![image](https://github.com/user-attachments/assets/d93a4556-f1c1-4ad3-ae61-72c50fe0dc01)