#4431 - Student View change request revamp - Application Submission (#4565)
## API-related changes
- Created a new endpoint `@Post(":applicationId/change-request")` to
allow the creation of the new application change request.
- Application Controller Refactor
- `validateApplicationSubmission` method created to centralize the
application submissions validations and get the expected application
data to be saved.
- Method `validateSubmitApplicationData` moved from the controller to
the controller service.
  
## Queue-consumers-related changes
- Adapted the method that selects the workflows to be started to also
consider edited applications with change requests.

## Web-related changes
- Adapted the current student application component to handle the
submission of a change request. There aren't clear content directions
right now, and the goal is to allow the submission of a change request,
keeping most of the validations in place.
- Created a new route speficic for the new appplication.
- The logic to display an application as read-only was adapted as a
sequence of `if/else` to improve readability. if there is a better way,
please share 😉
- Submitting a change request redirects to the application details with
the intention to show the tracker card to be implemented in the upcoming
PR. There are no directions on it right now.
- The UI is currently preventing the user from submitting a change
request when one is already in progress. No further UI direction has
been given so far.

### Current UI for change request


![image](https://github.com/user-attachments/assets/05d1874f-5487-40b8-8e94-a20994f5ddb7)

_Note:_ E2E tests are yet to be included.