#4489 - Formio Changes: Split FT/PT [Bug Fix] (#4592)
### As a part of this PR, the following bug was fixed:

**Bug:** When the institution location is selected in the dropdown for
the student application, the program component does not refresh properly
and still holds the previous (stale) data in the dropdown along with the
new data for the newly selected location.

**Fix:** To fix the above issue, I have added the below code snippet to
force remove the stale data from both the programs and offerings
dropdown before loading the fresh data.

**Code modification:** `formioUtils.setComponentValue(form,
PROGRAMS_DROPDOWN_KEY, "");`