#4431 - Student View change request revamp - Tracker Card (#4589)
## API

- Added extra information for the "completed application tracker card"
endpoint to allow the retrieval of the information for the change
requests: student income, supporting users data and income, and Ministry
approval.
- `ChangeRequestInProgressAPIOutDTO` added to extend the current DTO
reusing the same interfaces already used to report `Waiting Statuses`.
- Create a new endpoint to allow the student to cancel a change request.
This enpoint was created under the application route, but we can discuss
the possibility of moving it to a dedicated route (not in this PR).
Right now, the change request-related endpoints considered are:
  - Implemented in a previous PR
    - POST `students/application/:applicationId/change-request`
  - Implemented in this PR
    - PATCH `students/application/:applicationId/cancel-change-request`
  - To be implemented and not defined yet (Minstry approval)
    - PATCH `aest/application/:applicationId/change-request`

## Supporting Users

Allow the supporting users to submit data to a change request; before,
it prevented the data from being submitted to any application that was
not in progress.

## Workers

- Removed condition that prevents an assessment from being saved when
the application is not `in progress`.
- Changed `sequencial processing` method
`getOutstandingAssessmentsForStudentInSequence` to ignore edited
applications.

## Tracker Card States (Student View)

### Submitted


![image](https://github.com/user-attachments/assets/1ca75671-8d23-41a5-ab3a-ee3c7fadd064)

### Waiting for Ministry Approval


![image](https://github.com/user-attachments/assets/8728c592-0cfd-46e6-96f2-a246e94fd6a3)

### Waiting Partner Information


![image](https://github.com/user-attachments/assets/2148bab3-431f-4e6d-add8-68924f72ffca)+

### Waiting on Multiple Conditions (two parents where one provided their
data already)


![image](https://github.com/user-attachments/assets/0b9c4e8b-006c-4fba-b763-a08f03f62724)

### Ministy View with Disabled Buttons


![image](https://github.com/user-attachments/assets/d90aa5f2-ef69-42cd-a0da-db20543f272d)

### Validation for archived application

Added validation for archived application for the current and the new
change requests, extending it to the UI buttons.

### Some UI considerations

- "Student Aid" was adjusted later to "StudentAid" (without spaces).
- Opted for generating the list of waiting processes using `v-if`
conditions to allow the plain texts to be expanded in the future and
have elements other than plain text (e.g., links).
- Filter to restrict the version statuses displayed in the application
summary executed on the client side considering that will be likely to
filter zero to couple recods and adding the filter in the API was not
considered worth. Open for discussion, but if there is a strong desire
and Team agreement, I would prefer to take care in an upcoming PR.
- The redirect to the route
`StudentRoutesConst.DYNAMIC_FINANCIAL_CHANGE_REQUEST_APP_FORM` demands
extra information to be retrieved (program year and form name). A
refcator may be required but I am not sure if it would be in this ticket
scope.

#### Versions Expander Behavior Changed

The versions expander was conditionally displayed before based on the
fact that the application was the "Original" or not, which was not
correct since we also need to display, for instance, cancelled changed
requests. The behavior was changed to allow the button to always be
displayed and keep adding the versions on demand.


![image](https://github.com/user-attachments/assets/bb0adcf8-4691-4df8-83f3-997fa243bfbb)