#4431 - Student View change request revamp - Tracker Card API E2E tests (#4618)
Added API tests for the student endpoint that retrieve the tracker card
information.

ApplicationStudentsController(e2e)-getCompletedApplicationDetails
√ Should get application with change request details when there is an
in-progress change request pending Ministry approval.
√ Should get application with change request details including waiting
student income and waiting partner data when there is an in-progress
change request with a partner associated.
√ Should get application with change request details including with a
received student income, parent 1 waiting income, parent 2 waiting data
when there is an in-progress change request with parents associated.

_Note:_ Found an existing issue when there are two parents, and the
income records are not fully provided yet for both parents. If the
parent 2 provides its data and income, it will be consumed as belonging
to parent 1.