#4520 - Formio split part-1 (#4629)
# Formio split part-1

## DB Migration

- [x] Created new table `sims.dynamic_form_configurations` to configure
the form definition extraction of dynamic forms.
- [x] Added a constraint in the table to avoid configuration duplicates.
- [x] Inserted the dynamic form configuration values into the new table
based on existing values from `sims.program_years`.

## Rollback Evidence

![image](https://github.com/user-attachments/assets/69129d49-9469-4b04-998a-0fb1b65c8e32)

## Dynamic form configuration service

- [x] Created `DynamicFormConfigurationService` which is added to
`DynamicFormConfigurationModule` which is created as global module and
added to `AppModule`.
- [x] All the dynamic form configurations are loaded on start-up to
avoid going to DB every time when dynamic form is looked up. This
approach is done as these configurations are very rarely expected to be
modified. But if modified, it will require `restart rollout` for the
changes to take effect.