#4520 -  FormIO split final part - cleanup and backward compatibility (#4679)
# FormIO split final part

## Conditional variable injections

- [x] The formio hidden component value
`howWillYouBeAttendingTheProgram` is injected conditionally based on the
component being present in form. This component is present only until PY
2024-25.
- [x] The offering intensity is being displayed as read only text-input
in all the older forms until PY 2024-25. The value for this text input
is calculated(NOT calculated on server as it is not required) in the
form itself delegating the logic to the form.

![image](https://github.com/user-attachments/assets/7629e05d-4bc7-460a-bbea-e274dd690ab6)

## API Dry run submission backward compatibility
- [x] The API dry run submission was not failing when
`howWillYouBeAttendingTheProgram` is missing for forms until PY 2024-25
allowing the erroneous data to be saved. Hence updated the injection of
`howWillYouBeAttendingTheProgram` always on server side from the
application's offering intensity. Added a TODO comment there to remove
this injection once PY 2024-25 is set to be inactive.
## E2E Tests

- [x] Removed `howWillYouBeAttendingTheProgram` in the E2E tests
wherever it is applicable.
- [x] Added an E2E test for application submit for the latest active PY
to ensure that the newly configured program year has all required
configurations to submit applications.