#4690 - Update Deploy-All to use tag Git Reference - Build-ref Fallback (#4698)
Proposed solution to have the GitHub ref provided from both workflow
entries: `workflow_dispatch` and `workflow_call`.
When `workflow_call` is executed, it will be provided by the parent
workflow.
When `workflow_dispatch` is executed, it will use the `github.ref_name`
as the default value.
_Note:_
1 - The env variable is enough to execute the jobs, but it is not
accessible from the reusable workflows `with` context for
`deployCamundaDefinitions` and `deployFormioDefinitions`.
2 - Reference for the ternary operation used to set the ENV variable:
https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/evaluate-expressions-in-workflows-and-actions#example