#4317- Change request - Make program read only (#4706)
# Change Request - Make program section read-only

## Form
- [x] New warning banner has been added to the program tab.

![image](https://github.com/user-attachments/assets/ffea412a-a300-4f02-bea7-7845f03fb6ee)

- [x] The whole program section has been made read-only using the logic
in formio.

![image](https://github.com/user-attachments/assets/064ca8a1-95d0-4d7b-a242-b3e794257f1b)

- [x] Similar to the application read-only view, the drop down
components are hidden and instead displayed as read-only text labels in
program tab.

![image](https://github.com/user-attachments/assets/3fbdec44-9586-4d04-be34-e3f976a6131d)

- [x] To copy the program data for a change request from it's current
application, a new hidden variable component which is calculated at
server was added. `programPersistentProperties`

## API 

- [x] During the submission of application change request, all the
program persistent properties are copied from it's current application.
The list of persistent property names are kept in the student
application form as it is closely associated to the form.

- [x] A minor refactor was done to inject the following properties from
server side to avoid them to be manipulated.

`programYearStartDate` and `programYearEndDate`