#4617 - Ministry View Change Request Revamp (approval and denial) - E2E tests (#4718)
- Minor fix while saving the change request to persist the change date
(as pointed out in the previous PR).
- Modal refactor using the most recent practices to control modal
processing, loading, and possible errors.

## E2E Tests


ApplicationChangeRequestAESTController(e2e)-assessApplicationChangeRequest
√ Should approve a change request and copy the offering and appeal when
the application change request is waiting for approval.
√ Should approve a change request and copy the offering, and no appeals
when the application change request is waiting for approval, and no
appeals are present.
√ Should be able to decline a change request and create a student note
when the application change request is waiting for approval.
√ Should throw a BadRequestException when the application change request
approval has an invalid status.
√ Should throw UnprocessableEntityException when the application change
request was cancelled.
√ Should throw NotFoundException when the application change request
does not exist.