#4438 - MSFAA Sequence Group update (#4739)
## MSFAA Sequences update

### Release instructions

Note: Please coordinate with the business to set the life-time sequence
number as a part of release instructions for the Full-time and Part-time
sequence groups.

### MSFAA lifetime sequence

- [x] Based on the sequence group name
`MSFAA_${offeringIntensity}_SENT_FILE` updated the MSFAA header for PT
and FT to include a lifetime sequence number.

 Full time: 


![image](https://github.com/user-attachments/assets/d02bc138-c6be-4b6c-87b3-f046d0818758)


![image](https://github.com/user-attachments/assets/48cd750b-55d6-474f-8aad-9d5c0d8e66b9)


Part Time: 


![image](https://github.com/user-attachments/assets/0a93ddc7-8ec9-4ab7-be1b-db9a2b0b9184)


![image](https://github.com/user-attachments/assets/35c81a91-d7fc-4e81-a91c-132e08763129)

## MSFAA Daily filename sequence

- [x] Daily filename sequence is an existing sequence where a sequence
gap has been added to the daily file name sequence to avoid the conflict
with legacy. This sequence gap has been removed for Part-time as the
legacy MSFAA generation is no longer applicable to it. But the Full-time
MSFAA still has the sequence gap added.

Full time: 


![image](https://github.com/user-attachments/assets/ee925018-8697-4a90-8d2f-80a27728bff7)


![image](https://github.com/user-attachments/assets/c6e1cf72-c5a7-4927-b6e4-5c58340d6007)

Part time: 


![image](https://github.com/user-attachments/assets/b7655763-b07c-453c-805f-81919921aabd)


![image](https://github.com/user-attachments/assets/513c1c82-cd52-488a-b29c-624b553f6cb6)

## E2E Tests

- [x] Updated the PT E2E Test
- [x] Added a FT E2E Test